fix: add support for WASM targets via `libm`
