+++ packages/vexide-startup/build.rs
@@ -3,5 +3,4 @@
 fn main() {
     let manifest_dir = env!("CARGO_MANIFEST_DIR");
     println!("cargo:rustc-link-search=native={manifest_dir}/link");
-    println!("cargo:rustc-link-lib=static=m");
 }
